requirements.txt
# Core Dependencies
azure-identity==1.15.0
azure-ai-inference==1.0.0b4
openai==1.12.0

# Azure Services
azure-search-documents==11.4.0
azure-core==1.30.0

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
pandas==2.2.0
pillow==10.2.0

# Embeddings and Vector Store
sentence-transformers==2.3.1
numpy==1.26.3

# Agent Framework
langchain==0.1.6
langchain-community==0.0.19
langchain-openai==0.0.5

# OpenTelemetry
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-requests==0.43b0

# UI
streamlit==1.31.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3
pydantic==2.6.0
requests==2.31.0

# Evaluation
rouge-score==0.1.2
nltk==3.8.1

# Development
pytest==8.0.0
black==24.1.1
